amending logic around share bookings
